# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SkyBloomTD Edge API
  version: 0.1.0
  description: |
    Public API exposed through the Nginx reverse proxy.

    Only the reverse proxy publishes a host port. Backend services are private
    Docker-network services and should be accessed through these paths.

    Protected backend routes require a Supabase JWT stored in the
    `skybloom_access_token` cookie. Nginx verifies the cookie through
    `user-service /auth/verify` and forwards trusted identity headers to
    private services.
servers:
  - url: "{publicBaseUrl}"
    description: Public reverse proxy base URL
    variables:
      publicBaseUrl:
        default: http://your-server-domain
...
paths:
...
  /api/document-content/documents/{document_id}:
    delete:
      tags: [Document Content]
      summary: Delete a document and generated content
      description: |
        Deletes a document owned by the authenticated user. The service removes
        S3/local document assets, generated quizzes/levels for the document,
        document chunks/sections, and the document row.
      operationId: deleteDocument
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "204":
          description: Document and generated content were deleted.
        "400":
          description: Invalid document id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid authentication cookie.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Document was not found for the authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Asset deletion or database deletion failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/document-content/documents/{document_id}/chapters:
    get:
      tags: [Document Content]
      summary: List chapters for a document
      description: |
        Returns indexed chapters for a document owned by the authenticated user.
        A document that is not ready yet may return an empty chapter list.
      operationId: listDocumentChapters
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200":
          description: Document chapters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListChaptersResponse"
        "400":
          description: Invalid document id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid authentication cookie.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Document was not found for the authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Database lookup failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/document-content/chapters/{chapter_id}/sub-chapters:
    get:
      tags: [Document Content]
      summary: List sub-chapters for a chapter
      description: Returns indexed sub-chapters for a chapter owned by the authenticated user.
      operationId: listChapterSubChapters
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/ChapterId"
      responses:
        "200":
          description: Chapter sub-chapters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListSubChaptersResponse"
        "400":
          description: Invalid chapter id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid authentication cookie.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Chapter was not found for the authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Database lookup failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: skybloom_access_token
  parameters:
...
    DocumentId:
      name: document_id
      in: path
      required: true
      description: Document id returned by upload or document listing.
      schema:
        type: string
        format: uuid
    ChapterId:
      name: chapter_id
      in: path
      required: true
      description: Chapter id returned by the document chapters endpoint.
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: authentication required
...
    ListChaptersResponse:
      type: object
      required:
        - chapters
      properties:
        chapters:
          type: array
          items:
            $ref: "#/components/schemas/ChapterSummary"
    ChapterSummary:
      type: object
      required:
        - chapter_id
        - document_id
        - chapter_index
        - title
        - start_line
        - end_line
        - created_at
      properties:
        chapter_id:
          type: string
          format: uuid
        document_id:
          type: string
          format: uuid
        chapter_index:
          type:
            - integer
            - "null"
        title:
          type:
            - string
            - "null"
          example: Linear transformations
        start_line:
          type:
            - integer
            - "null"
        end_line:
          type:
            - integer
            - "null"
        created_at:
          type: string
          format: date-time
    ListSubChaptersResponse:
      type: object
      required:
        - sub_chapters
      properties:
        sub_chapters:
          type: array
          items:
            $ref: "#/components/schemas/SubChapterSummary"
    SubChapterSummary:
      type: object
      required:
        - sub_chapter_id
        - document_id
        - chapter_id
        - sub_chapter_index
        - title
        - start_line
        - end_line
        - created_at
      properties:
        sub_chapter_id:
          type: string
          format: uuid
        document_id:
          type: string
          format: uuid
        chapter_id:
          type: string
          format: uuid
        sub_chapter_index:
          type:
            - integer
            - "null"
        title:
          type:
            - string
            - "null"
          example: Matrix representation
        start_line:
          type:
            - integer
            - "null"
        end_line:
          type:
            - integer
            - "null"
        created_at:
          type: string
          format: date-time
...